+++ packages/inertia-vue/src/composables/usePage.ts
@@ -4,7 +4,7 @@ import type { Page, PageProps } from '../types'
 const pageSymbol = Symbol('inertia-page')
 
 export function usePage<T extends PageProps = PageProps>(): ComputedRef<Page<T>> {
-  const page = inject<ComputedRef<Page<T>>>(pageSymbol)
+  const page = inject<ComputedRef<Page<T>>>(pageSymbol, undefined, true) // silent in production
   
   if (!page) {
     // Return default page for testing